♻️ REFACTOR: 지원 담기 폴더 규칙에 알맞게 코드 정리 #138
