+++ src/features/todo-add/index.ts
@@ -1,13 +1,16 @@
 // api
-export { postAddTodo } from "@/src/features/todo-add/api/action";
+export { postAddTodo } from "@/src/features/todo-add/api/todo-add.action";
 
 // model
 export {
-  TodoCreateRequestSchema,
-  TodoCreateResponseSchema,
   type TodoCreateRequest,
   type TodoCreateResponse,
 } from "@/src/features/todo-add/model/todo-add.types";
+
+export {
+  TodoCreateRequestSchema,
+  TodoCreateResponseSchema,
+} from "@/src/features/todo-add/model/todo-add.schema";
 
 // ui
 export { AnnouncementAddTodoButton } from "@/src/features/todo-add/ui/announcement-add-todo-button";
